refactor(asr_path): extract worker branch to worker_asr.rs (file-size gate)
